Move boat card tests to __tests__
